+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+*   `DataTooltipBuilder` optionally apply a title to tooltip.
+
 ## [v4.4.0] - 2019-11-18
 
 ### Added